requirements.txt
discord.py
python-dotenv~=1.0.0
elasticsearch>=7.8.0
motor~=3.3.1
requests~=2.31.0
dnspython
python-dateutil~=2.8.2
algoliasearch>=2.0,<3.0
asyncio>=3.4,<4.0
aiohttp>=2.0,<4.0
async_timeout>=2.0,<4.0
fuzzywuzzy[speedup]~=0.18.0
cryptography~=41.0.3
aiofiles==22.1.0
msgpack~=1.0.5
